added frontened UI

+++ backend/controllers/postController.js
@@ -1,7 +1,14 @@
 const asyncHandler = require('express-async-handler')
 const Post = require('../models/postModal')
+const User = require('../models/userModal')
+
+// const getAllPosts = asyncHandler(async (req, res) => {
+//     const posts = await Post.find().populate('user', 'name email profilePic');
+//     res.status(200).json(posts);
+// });
+
 const getPosts = asyncHandler(async(req,res) =>{
-    const posts = await Post.find()
+    const posts = await Post.find({user: req.user.id})
     res.status(200).json(posts)
 }
 )
@@ -10,7 +17,9 @@ const setPosts = asyncHandler(async(req,res) =>{
         res.status(400).json( {message: 'please enter post'})
     }
      const post = await Post.create({
-         text : req.body.text
+         text : req.body.text,
+         user: req.user.id,
+         image: req.body.text,
      })
     res.status(200).json( post)
 })
@@ -19,6 +28,15 @@ const updatePosts = asyncHandler(async(req,res) =>{
     if(!post){
         res.status(400)
         throw new Error('Post description not found')
+    }
+    const user = await User.findById(req.user.id)
+    if(!user){
+        res.status(401)
+        throw new Error ('user not found')
+    }
+    if(post.user.toString() != user.id){
+        res.status(401)
+        throw new Error ('user not authorised')
     }
     const updatePost = await Post.findByIdAndUpdate(req.params.id,req.body,{
         new:true,
@@ -31,8 +49,17 @@ const deletePosts = asyncHandler(async(req,res) =>{
         res.status(400)
         throw new Error('Post description not found')
     }
+    const user = await User.findById(req.user.id)
+    if(!user){
+        res.status(401)
+        throw new Error ('user not found')
+    }
+    if(post.user.toString() != user.id){
+        res.status(401)
+        throw new Error ('user not authorised')
+    }
     await post.deleteOne()
-    res.status(200).json( {message: `delete post ${req.params.id}`})
+    res.status(200).json( {id: req.params.id})
 })
 module.exports = {
     getPosts,setPosts,updatePosts,deletePosts,
